Fix mean airtime

+++ NYC-Flights13-Backend/Models/AirtimeAtOriginDTO.cs
@@ -8,7 +8,7 @@ namespace NYC_Flights13_Backend.Models
 {
     public class AirtimeAtOriginDTO
     {
-        public int AirTime { get; set; }
+        public float AirTime { get; set; }
 
         public string Origin { get; set; }
